refactor: history details
refactor history details so that it is a standalone component, move schemas under uiSchema/history for future history schemas
